realm: Add storage in deployment (#205)
This adds another after deployment script for Waypoint to run, it amends
the ECS task definition with volume and mount information. It removes
the update-service call in the tagging script to avoid updating the
service twice.

The deployment startup commands now refer to the persistent filesystem
which is mounted at /persistent.