host: Add more realm awareness to code mode (#695)
This hoists the openFile resource from CodeMode component into
OperatorModeStateService so it and other services can be more
realm-aware via a realmURL getter derived from openFile. Previously the
active file would not reliably be highlighted in the file tree when
viewing a non-“default” realm, nor would recent files always be tracked
across realms.

This necessitated changing openDirs to be realm-aware, so it’s now a
mapping of realm URL to open directories, with deserialisation to turn
it into a TrackedMap of TrackedArrays. It also required adding a
file-change subscription handler to OperatorModeStateService to enable
state management around 404 rendering and dismissal.