Testing monorepo without dependenciesMeta.injected
I'm hitting places where typescript doesn't understand this and leaves holes in the type analysis. We added dependenciesMeta.injected originally so that conflicting peer dependencies would be resolved correctly, but if we don't *have* any conflicting peerDependencies now this might work fine.

Alternatively, to avoid the type-analysis failures we would need skipLibCheck to be off.