Refactor MissingEmbeddedTemplate
We started refactoring this in preparation to remove several of the arguments to card components. We can't yet remove `@cardOrField` but this is still a nice small cleanup.

Along the way we realized we're not using very good types for card components throughout the system. Many are using the weakly-typed `BaseDefComponent` instead of `SignatureFor<>`, and the two actually disagree and have been allowed to diverge.