WIP
- [ ] make computeds run all the time by updating getter
- [ ] make relationshipMeta always return a descriptive value, and make field accessor never throw NotLoaded

Note from LM: The issue I have run into is that if the field accessor doesn't throw, there is no way for the indexer to wait for the card to be fully loaded (currently we are catching NotLoaded errors when attempting to render a card for indexing)

ef4: makes sense. we can make the relationship loading process register itself somehow. It's not unlike test waiters or fastboot deferred.