Avoid getLoaderFor in base/card-api
It's not necessary to use getLoaderFor anywhere in `base` because all of `base` is always loaded by an instance of our Loader, and that is the relevant instance to use for everything.

This fixes a crash when loading a linksTo a Card that can have more than one CodeRef depending on the specific timing of how it has already been accessed.