Fix bug where auto-attachment resource errors when new topmost card is saving
When a new card is being saved, a card resource is in a state where it can't provide card. This caused our AutoAttachment resource code to crash as it assumed that every StackItem could provide a card via its CardResrouce. This PR makes the AutoAttachment resource wait for all stack items to be ready before performing its logic, using an ember-concurrency restartableTask