Remove portion of test from matrix suite and add to host tests.
The Matrix suite had an AI Assistant test that mutates realm state, which is not desired in that suite. This commit removes that portion of the matrix suite test and reproduces the test in the host suite. Making it work required making the MockMatrixService a bit more faithful to the actual one.