Add another clientRequestId prefix
This fixes the “can patch code when there are multiple
patches” test. The file resource was ignoring the
invalidation because it had the source prefix, which
was stored in the known list, as it was produced by
this client.

Another approach would be to not include the clientRequestId
in the fetch request, because it was produced indirectly
via a bot patch, not directly through the code editor.