avoid racing destructors
This was previously relying on destructor ordering. The destructor for CodeEditor needed to run before the destructor for the monaco modifier, or otherwise it would not see any monaco content to save.

As the order isn't guaranteed and bugfixes changed it in the @glimmer/component 2.0 upgrade, I'm switching here to an implementation that has no race. The modifier itself offers a hook for doing something before the editor is disposed.