Apply consistent-type-imports rule
The eslint config differed between TS and GTS files, and only one had the consistent-type-imports rule (which we want to make `@embroider/vite` work better).

This unifies the config so the superset of all the rules we had configured applies to both TS/JS and GTS/GJS, and auto-fixes the resulting failures.